+++ layer/vulkan_reflex_layer.cpp
@@ -32,7 +32,7 @@
 // hopefully nobody is going to build this for 32-bit ARM Android
 #define VKAPI_ATTR
 
-#include "vkroots.h"
+#include <vkroots.h>
 
 #define LOG_CHANNEL "vkreflex_layer"
 #include "log.h"
