+++ src/meson.build
@@ -37,7 +37,7 @@ nvapi_dll = shared_library(target_name, [ nvapi_src, dxvk_nvapi_version ],
   cpp_args            : '-DDXVK_NVAPI_TARGET_NAME="'+target_name+'"',
   name_prefix         : '',
   dependencies        : [ lib_dxgi, lib_d3d11, lib_version ],
-  include_directories : [ nvapi_headers, vk_headers, dx_headers ],
+  include_directories : [ nvapi_headers, vk_headers, dx_headers, nvml_headers, nvofapi_headers ],
   vs_module_defs      : target_name+'.def',
   install             : true)
 
@@ -66,7 +66,7 @@ if dxvk_cpu_family == 'x86_64'
     cpp_args            : '-DDXVK_NVAPI_TARGET_NAME="'+target_name+'"',
     name_prefix         : '',
     dependencies        : [ lib_version ],
-    include_directories : [ nvapi_headers, vk_headers, dx_headers ],
+    include_directories : [ nvapi_headers, vk_headers, dx_headers, nvml_headers, nvofapi_headers ],
     vs_module_defs      : target_name+'.def',
     install             : true)
 endif
