+++ src/nvofapi_private.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "./nvapi_private.h"
-#include "../inc/nvofapi/nvOpticalFlowD3D11.h"
-#include "../inc/nvofapi/nvOpticalFlowD3D12.h"
-#include "../inc/nvofapi/nvOpticalFlowVulkan.h"
+
+#include <nvOpticalFlowD3D11.h>
+#include <nvOpticalFlowD3D12.h>
+#include <nvOpticalFlowVulkan.h>
